+++ AntonBabunin/src/main/java/ua/artcode/simplehttpserver/SimpleHttpServer.java
@@ -1,7 +1,8 @@
 package ua.artcode.simplehttpserver;
 
 import com.sun.net.httpserver.HttpServer;
-import ua.artcode.simplehttpserver.hoslders.employee.HandlerHolder;
+import ua.artcode.simplehttpserver.hoslders.HandlerHolder;
+import ua.artcode.simplehttpserver.hoslders.employee.*;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
@@ -14,12 +15,12 @@ public class SimpleHttpServer implements Runnable {
         try {
             server = HttpServer.create(new InetSocketAddress(8000), 0);
             server.createContext("/", new HandlerHolder());
-            server.createContext("/login", new HandlerHolder.HandlerLoginPost());
-            server.createContext("/employee", new HandlerHolder.HandlerEmployee());
-            server.createContext("/employee/createbill", new HandlerHolder.HandlerCreateBill());
-            server.createContext("/employee/addproduct", new HandlerHolder.HandlerAddProduct());
-            server.createContext("/employee/getbill", new HandlerHolder.HandlerGetBill());
-            server.createContext("/employee/bills", new HandlerHolder.HandlerGetBills());
+            server.createContext("/login", new HandlerLoginPost());
+            server.createContext("/employee", new HandlerEmployee());
+            server.createContext("/employee/createbill", new HandlerCreateBill());
+            server.createContext("/employee/addproduct", new HandlerAddProduct());
+            server.createContext("/employee/getbill", new HandlerGetBill());
+//            server.createContext("/employee/bills", new HandlerHolder.HandlerGetBills());
 
 
         } catch (IOException e) {
