+++ AntonBabunin/src/main/java/ua/artcode/simplehttpserver/hoslders/employee/HandlerAddProduct.java
@@ -1,0 +1,16 @@
+package ua.artcode.simplehttpserver.hoslders.employee;
+
+import com.sun.net.httpserver.HttpExchange;
+import ua.artcode.simplehttpserver.hoslders.HandlerHolder;
+
+import java.io.IOException;
+
+public class HandlerAddProduct extends HandlerHolder {
+    public HandlerAddProduct() throws IOException {
+    }
+
+    @Override
+    public void handle(HttpExchange httpExchange) throws IOException {
+
+    }
+}
