+++ AntonBabunin/src/main/java/ua/artcode/simplehttpserver/hoslders/employee/HandlerCreateBill.java
@@ -1,0 +1,16 @@
+package ua.artcode.simplehttpserver.hoslders.employee;
+
+import com.sun.net.httpserver.HttpExchange;
+import ua.artcode.simplehttpserver.hoslders.HandlerHolder;
+
+import java.io.IOException;
+
+public class HandlerCreateBill extends HandlerHolder {
+    public HandlerCreateBill() throws IOException {
+    }
+    @Override
+    public void handle(HttpExchange httpExchange) throws IOException {
+
+    }
+
+}
